+++ tests/test_examples.py
@@ -151,9 +151,10 @@ async def test_libreoffice_example(path_tmp: Path, name: str):
     [
         "compile_typst_args",
         "compile_typst_dep",
-        pytest.param("compile_typst_dep_error", marks=pytest.mark.xfail),
+        "compile_typst_dep_error",
         "compile_typst_error",
         "compile_typst_external",
+        "compile_typst_html",
         "compile_typst_png",
         "compile_typst_png_multi",
         "compile_typst_relpath",